router/requirements.txt
kafka-python==2.0.2
pika==1.3.2
tenacity==8.2.3
pyyaml==6.0.1
